refactor(renderer): extract remote command helpers (PR-renderer-6)
Surgically re-applied on top of main after PR-renderer-7 (usePairingFlow)
merged first (concurrent wave conflict). All same changes as the original
commit:
  - useRemoteSession hook owns busy state + command queue + flushQueuedCommands
  - KEYBOARD_COMMAND_MAP exported from lib/remoteCommands.ts
  - 5 inline queue functions removed from App.tsx
  - CommandDispatchRequest/RemoteCommand/RemoteCommandSource type imports removed
  - keyboardCommandMap constant removed from App.tsx

App.tsx: 1980 -> 1867 LOC (-113)
